Fix occasional reload triggers by temp files
This is caused by a race condition: extension creates the files and adds it to temp paths, watcher notices this, extension gets done with the temp file and removes both the file and the path from temp paths, and watcher just now triggers the loadOnChange with a file that does not exist anymore, thus passing the condition. Existence check is a sufficient remedy.